Move Position to its own file
